backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2

langgraph
langchain-core
openai
gTTS
whisper
ffmpeg-python
openai-whisper
langchain_google_genai
pydantic-settings
pymongo
python-multipart
cloudinary